# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateHomepageSectionRequest:
    properties:
      content: {}
      description:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      section_key:
        type: string
      sort_order:
        type: integer
      subtitle:
        type: string
      title:
        type: string
    required:
    - section_key
    type: object
  dto.CreateSEOPageRequest:
    properties:
      canonical_url:
        type: string
      meta_description:
        type: string
      meta_keywords:
        type: string
      meta_title:
        type: string
      og_description:
        type: string
      og_image:
        type: string
      og_title:
        type: string
      page_slug:
        type: string
    required:
    - page_slug
    type: object
...
  dto.LoanInquiryRequest:
    properties:
      city:
        type: string
      email:
        type: string
      full_name:
        type: string
      monthly_income:
        type: string
      phone:
        type: string
      service:
        type: string
    required:
    - city
    - email
    - full_name
    - monthly_income
    - phone
    - service
    type: object
...
  dto.UpdateHomepageSectionRequest:
    properties:
      content: {}
      description:
        type: string
      image:
        type: string
      is_active:
        type: boolean
      section_key:
        type: string
      sort_order:
        type: integer
      subtitle:
        type: string
      title:
        type: string
    type: object
  dto.UpdateInquiryStatusRequest:
    properties:
      status:
        type: string
    required:
    - status
    type: object
  dto.UpdateSEOPageRequest:
    properties:
      canonical_url:
        type: string
      meta_description:
        type: string
      meta_keywords:
        type: string
      meta_title:
        type: string
      og_description:
        type: string
      og_image:
        type: string
      og_title:
        type: string
      page_slug:
        type: string
    type: object
...
  models.HomepageSection:
    properties:
      content: {}
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image:
        type: string
      is_active:
        type: boolean
      section_key:
        type: string
      sort_order:
        type: integer
      subtitle:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  models.LoanInquiry:
    properties:
      city:
        type: string
      created_at:
        type: string
      email:
        type: string
      full_name:
        type: string
      id:
        type: integer
      monthly_income:
        type: string
      phone:
        type: string
      service:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  models.SEOPage:
    properties:
      canonical_url:
        type: string
      created_at:
        type: string
      id:
        type: integer
      meta_description:
        type: string
      meta_keywords:
        type: string
      meta_title:
        type: string
      og_description:
        type: string
      og_image:
        type: string
      og_title:
        type: string
      page_slug:
        type: string
      updated_at:
        type: string
    type: object
...
  utils.APIResponse:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
  utils.PaginatedResponse:
    properties:
      data: {}
      page:
        type: integer
      page_size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
      total_pages:
        type: integer
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API server for GK Capital finance website.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GK Capital API Documentation
  version: "1.0"
paths:
...
  /api/admin/homepage:
    get:
      description: Get paginated list of all homepage sections
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get all homepage sections for administration
      tags:
      - Admin Homepage Sections
    post:
      consumes:
      - application/json
      description: Create a homepage section
      parameters:
      - description: Create Homepage Section Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateHomepageSectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.HomepageSection'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new homepage section
      tags:
      - Admin Homepage Sections
  /api/admin/homepage/{id}:
    delete:
      description: Delete a homepage section by ID
      parameters:
      - description: Homepage Section ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete a homepage section
      tags:
      - Admin Homepage Sections
    put:
      consumes:
      - application/json
      description: Update details of a homepage section by ID
      parameters:
      - description: Homepage Section ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Homepage Section Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateHomepageSectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.HomepageSection'
              type: object
      security:
      - BearerAuth: []
      summary: Update an existing homepage section
      tags:
      - Admin Homepage Sections
  /api/admin/loan-inquiries:
    get:
      description: Get list of loan applications with search and status filters
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by status (new/contacted/approved/rejected)
        in: query
        name: status
        type: string
      - description: Search query (name, email, phone, city)
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.PaginatedResponse'
      security:
      - BearerAuth: []
      summary: Get loan inquiries
      tags:
      - Admin Inquiries
  /api/admin/loan-inquiries/{id}:
    patch:
      consumes:
      - application/json
      description: Update status of an application by ID
      parameters:
      - description: Inquiry ID
        in: path
        name: id
        required: true
        type: integer
      - description: New status request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateInquiryStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Update loan inquiry status
      tags:
      - Admin Inquiries
  /api/admin/seo:
    get:
      description: Get list of all SEO entries for management
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.SEOPage'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Get all SEO page configurations
      tags:
      - Admin SEO Pages
    post:
      consumes:
      - application/json
      description: Setup SEO meta tags for a page slug
      parameters:
      - description: Create SEO configuration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSEOPageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SEOPage'
              type: object
      security:
      - BearerAuth: []
      summary: Create a new SEO configuration
      tags:
      - Admin SEO Pages
  /api/admin/seo/{id}:
    delete:
      description: Delete an SEO configuration by ID
      parameters:
      - description: SEO Configuration ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/utils.APIResponse'
      security:
      - BearerAuth: []
      summary: Delete an SEO configuration
      tags:
      - Admin SEO Pages
    put:
      consumes:
      - application/json
      description: Update SEO parameters by ID
      parameters:
      - description: SEO Configuration ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update SEO configuration request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSEOPageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SEOPage'
              type: object
      security:
      - BearerAuth: []
      summary: Update an existing SEO configuration
      tags:
      - Admin SEO Pages
...
  /api/loan-inquiry:
    post:
      consumes:
      - application/json
      description: Submit the loan application form
      parameters:
      - description: Loan inquiry application data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoanInquiryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.LoanInquiry'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Submit a loan inquiry (Apply Now)
      tags:
      - Loan Inquiry
  /api/seo:
    get:
      description: Fetch meta tags, canonical URL and open graph fields for a public
        page
      parameters:
      - description: Page slug (e.g. home, about, services, contact)
        in: query
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/utils.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.SEOPage'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.APIResponse'
      summary: Get SEO configurations by page slug
      tags:
      - Public Content
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"